+++ deno.json
@@ -1,6 +1,6 @@
 {
   "name": "@nathanwhit/deno-shell-setup",
-  "version": "0.5.0",
+  "version": "0.6.0",
   "exports": "./src/main.ts",
   "imports": {
     "@david/which": "jsr:@david/which@^0.4.1",
